fix: resolve "vite" relative to the project root instead of using a dependency, if possible
