+++ packages/vitest/src/node/ast-collect.ts
@@ -16,10 +16,10 @@ import {
 import { unique } from '@vitest/utils/helpers'
 import { ancestor as walkAst } from 'acorn-walk'
 import { relative } from 'pathe'
-import { parseAst } from 'vite'
 import { createIndexLocationsMap } from '../utils/base'
 import { createDebugger } from '../utils/debugger'
 import { detectCodeBlock } from '../utils/test-helpers'
+import { parseAst } from '../vite'
 
 interface ParsedFile extends File {
   start: number
