+++ packages/vitest/src/node/coverage.ts
@@ -1,9 +1,9 @@
 import type { CoverageMap } from 'istanbul-lib-coverage'
-import type { TransformResult } from 'vite'
 import type { Vitest } from '../node/core'
 import type { BaseCoverageOptions, CoverageModuleLoader, CoverageProvider, ReportContext, ResolvedCoverageOptions } from '../node/types/coverage'
 import type { SerializedCoverageConfig } from '../runtime/config'
 import type { AfterSuiteRunMeta } from '../types/general'
+import type { TransformResult } from '../vite'
 import type { TestProject } from './project'
 import { existsSync, promises as fs, readdirSync, writeFileSync } from 'node:fs'
 import module from 'node:module'
