+++ packages/vitest/src/node/environments/fetchModule.ts
@@ -1,8 +1,8 @@
 import type { Span } from '@opentelemetry/api'
-import type { DevEnvironment, EnvironmentModuleNode, FetchResult, Rollup, TransformResult } from 'vite'
 import type { FetchFunctionOptions } from 'vite/module-runner'
 import type { FetchCachedFileSystemResult } from '../../types/general'
 import type { OTELCarrier, Traces } from '../../utils/traces'
+import type { DevEnvironment, EnvironmentModuleNode, FetchResult, Rollup, TransformResult } from '../../vite'
 import type { FileSystemModuleCache } from '../cache/fsModuleCache'
 import type { VitestResolver } from '../resolver'
 import type { ResolvedConfig } from '../types/config'
@@ -10,7 +10,7 @@ import { existsSync, mkdirSync } from 'node:fs'
 import { readFile } from 'node:fs/promises'
 import { isExternalUrl, unwrapId } from '@vitest/utils/helpers'
 import { join } from 'pathe'
-import { fetchModule } from 'vite'
+import { fetchModule } from '../../vite'
 import { hash } from '../hash'
 import { normalizeResolvedIdToUrl } from './normalizeUrl'
