+++ packages/vitest/src/node/plugins/coverageTransform.ts
@@ -1,4 +1,4 @@
-import type { Plugin as VitePlugin } from 'vite'
+import type { Plugin as VitePlugin } from '../../vite'
 import type { Vitest } from '../core'
 
 export function CoverageTransform(ctx: Vitest): VitePlugin {
