+++ packages/vitest/src/node/plugins/metaEnvReplacer.ts
@@ -1,4 +1,4 @@
-import type { Plugin } from 'vite'
+import type { Plugin } from '../../vite'
 import { cleanUrl } from '@vitest/utils/helpers'
 import MagicString from 'magic-string'
 import { stripLiteral } from 'strip-literal'
