+++ packages/vitest/src/node/plugins/publicConfig.ts
@@ -1,13 +1,13 @@
 import type {
   ResolvedConfig as ResolvedViteConfig,
   UserConfig as ViteUserConfig,
-} from 'vite'
+} from '../../vite'
 import type { ResolvedConfig, UserConfig } from '../types/config'
 import { deepClone, slash } from '@vitest/utils/helpers'
 import * as find from 'empathic/find'
 import { resolve } from 'pathe'
-import { mergeConfig, resolveConfig as resolveViteConfig } from 'vite'
 import { configFiles } from '../../constants'
+import { mergeConfig, resolveConfig as resolveViteConfig } from '../../vite'
 import { resolveConfig as resolveVitestConfig } from '../config/resolveConfig'
 import { Vitest } from '../core'
 import { VitestPlugin } from './index'
