+++ packages/vitest/src/node/pool.ts
@@ -8,11 +8,11 @@ import type { BuiltinPool, ResolvedConfig } from './types/config'
 import * as nodeos from 'node:os'
 import { isatty } from 'node:tty'
 import { resolve } from 'pathe'
-import { version as viteVersion } from 'vite'
 import { rootDir } from '../paths'
 import { isWindows } from '../utils/env'
 import { getWorkerMemoryLimit, stringToBytes } from '../utils/memory-limit'
 import { getSpecificationsOptions } from '../utils/test-helpers'
+import { version as viteVersion } from '../vite'
 import { createBrowserPool } from './pools/browser'
 import { Pool } from './pools/pool'
