+++ packages/vitest/src/node/projects/resolveProjects.ts
@@ -13,8 +13,8 @@ import { limitConcurrency } from '@vitest/runner/utils'
 import { deepClone } from '@vitest/utils/helpers'
 import { basename, dirname, relative, resolve } from 'pathe'
 import { glob, isDynamicPattern } from 'tinyglobby'
-import { mergeConfig } from 'vite'
 import { configFiles as defaultConfigFiles } from '../../constants'
+import { mergeConfig } from '../../vite'
 import { VitestFilteredOutProjectError } from '../errors'
 import { initializeProject, TestProject } from '../project'
