+++ packages/vitest/src/node/types/browser.ts
@@ -2,10 +2,10 @@ import type { MockedModule } from '@vitest/mocker'
 import type { CancelReason } from '@vitest/runner'
 import type { Awaitable, ParsedStack, TestError } from '@vitest/utils'
 import type { StackTraceParserOptions } from '@vitest/utils/source-map'
-import type { Plugin, ViteDevServer } from 'vite'
 import type { BrowserCommands } from 'vitest/browser'
 import type { BrowserTraceViewMode } from '../../runtime/config'
 import type { BrowserTesterOptions } from '../../types/browser'
+import type { Plugin, ViteDevServer } from '../../vite'
 import type { TestProject } from '../project'
 import type { ApiConfig, ProjectConfig } from './config'
